move contents of __ini__.py to menus.py
